+++ CHANGELOG.md
@@ -10,6 +10,8 @@ v1.0.0 (in development)
 - Drop support for Python 3.6
 - `IndexPage`, `ProjectPage`, `DistributionPackage`, and `Link` have been
   changed from NamedTuples to dataclasses
+- Replaced `DistributionPackage.yanked` with separate `is_yanked` and
+  `yanked_reason` attributes
 
 v0.10.0 (2022-06-30)
 --------------------